Changed the downloading of docs back to a Gulp task, fixed downloading of docs so that it works if they have not been downloaded before and if they have

+++ src/Documentation/DocumentationDownloader.php
@@ -53,15 +53,18 @@ final class DocumentationDownloader
         foreach ($this->branches as $branch) {
             $rawDocsPath = "{$this->clonedDocPath}/$branch";
 
-            /**
-             * When cloning from GitHub, some files in the .git directory are read-only, which means we cannot delete
-             * them using normal PHP commands.  So, we first chmod all the files, then delete them.
-             */
-            foreach ($this->files->getFiles($rawDocsPath, true) as $file) {
-                chmod($file, 0777);
+            if ($this->files->exists($rawDocsPath)) {
+                /**
+                 * When cloning from GitHub, some files in the .git directory are read-only, which means we cannot delete
+                 * them using normal PHP commands.  So, we first chmod all the files, then delete them.
+                 */
+                foreach ($this->files->getFiles($rawDocsPath, true) as $file) {
+                    chmod($file, 0777);
+                }
+
+                $this->files->deleteDirectory($rawDocsPath);
             }
 
-            $this->files->deleteDirectory($rawDocsPath);
             $this->files->makeDirectory($rawDocsPath);
 
             // Clone the branch from GitHub into our temporary directory
